Interview_Ace/requirements.txt
PyMuPDF
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
werkzeug==2.3.7
google-generativeai==0.3.2
python-dotenv==1.0.0
gunicorn